refactor: Improve config UI with better separation of concerns
- Fix initial validation display issue (no ✓ Valid on page load)
- Separate CSS and JavaScript into external files
- Move getConfigDescription to repositories/config.ts for better organization
- Fix placeholder text truncation and value quoting issues
- Maintain existing functionality while improving code structure
